refac: restructured folders
