+++ lib/screens/runs/widgets/all_runs.dart
@@ -5,8 +5,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:logmap/models/run_model.dart';
 import 'package:logmap/models/truck_model.dart';
 import 'package:logmap/providers/driver_select_provider.dart';
-import 'package:logmap/shared/calculate_run_interval.dart';
-import 'package:logmap/shared/get_current_date.dart';
+import 'package:logmap/shared/functions/calculate_run_interval.dart';
+import 'package:logmap/shared/functions/get_current_date.dart';
 
 class AllRuns extends ConsumerWidget {
   const AllRuns({Key? key}) : super(key: key);
