--- lib/shared/custom_button.dart
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CustomButton extends StatelessWidget {
-  const CustomButton({
-    Key? key,
-    required this.onTap,
-    required this.text,
-  }) : super(key: key);
-  final String text;
-  final VoidCallback onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      style: ElevatedButton.styleFrom(
-        minimumSize: const Size(double.infinity, 40),
-      ),
-      onPressed: onTap,
-      child: Text(text),
-    );
-  }
-}
