+++ lib/shared/functions/calculate_run_interval.dart
