+++ lib/shared/functions/get_current_date.dart
